rename instagorillamux to instamux
