+++ .circleci/config.yml
@@ -76,14 +76,14 @@ workflows:
           name: "go1.11"
           go_version: "1.11"
           go111module: "on"
-          exclude_dirs: "./example/gorillamux ./instrumentation/instagorillamux"
+          exclude_dirs: "./example/gorillamux ./instrumentation/instamux"
       - build:
           name: "go1.10"
           go_version: "1.10"
           vendor_deps: "yes"
-          exclude_dirs: "./instrumentation/cloud.google.com/go ./example/gin  ./instrumentation/instagin ./example/gorillamux ./instrumentation/instagorillamux"
+          exclude_dirs: "./instrumentation/cloud.google.com/go ./example/gin  ./instrumentation/instagin ./example/gorillamux ./instrumentation/instamux"
       - build:
           name: "go1.9"
           go_version: "1.9"
           vendor_deps: "yes"
-          exclude_dirs: "./instrumentation/cloud.google.com/go ./instrumentation/instagin ./example/http-database-greeter ./example/grpc-client-server ./example/grpc-client-server/pb ./example/gin ./example/gorillamux ./instrumentation/instagorillamux"
+          exclude_dirs: "./instrumentation/cloud.google.com/go ./instrumentation/instagin ./example/http-database-greeter ./example/grpc-client-server ./example/grpc-client-server/pb ./example/gin ./example/gorillamux ./instrumentation/instamux"
